Separated appliance to its own class
